Rename to mapKernelInterfaces
